+++ app.py
@@ -1,112 +1,122 @@
-from contextlib import asynccontextmanager
+# from contextlib import asynccontextmanager
 
-from lihil import HTTPException, Json, Lihil, Payload, Resp, Route, Stream, Text, status
-from lihil.lihil import AppState
-from lihil.problems import HTTPException
-
-
-class Unhappiness(Payload):
-    scale: int
-    is_mad: bool
+# from lihil import HTTPException, Json, Lihil, Payload, Resp, Route, Stream, Text, status
+# from lihil.lihil import AppState
+# from lihil.problems import HTTPException
 
 
-class UserNotHappyError(HTTPException[Unhappiness]):
-    "user is not happy with what you are doing"
+# class Unhappiness(Payload):
+#     scale: int
+#     is_mad: bool
 
 
-class VioletsAreBlue(HTTPException[str]):
-    "how about you?"
-
-    __status__ = 418
+# class UserNotHappyError(HTTPException[Unhappiness]):
+#     "user is not happy with what you are doing"
 
 
-class UserNotFoundError(HTTPException[str]):
-    "Unable to find user with given user_id"
+# class VioletsAreBlue(HTTPException[str]):
+#     "how about you?"
 
-    __status__ = 404
-
-    ...
+#     __status__ = 418
 
 
-class User(Payload, kw_only=True, tag=True):
-    id: int
-    name: str
-    email: str
+# class UserNotFoundError(HTTPException[str]):
+#     "Unable to find user with given user_id"
+
+#     __status__ = 404
+
+#     ...
 
 
-class Order(Payload, tag=True):
-    order_id: str
-    price: float
+# class User(Payload, kw_only=True, tag=True):
+#     id: int
+#     name: str
+#     email: str
 
 
-rusers = Route("users")
+# class Order(Payload, tag=True):
+#     order_id: str
+#     price: float
 
 
-class MyState(AppState): ...
+# rusers = Route("users")
 
 
-@asynccontextmanager
-async def lifespan(app: Lihil[MyState]):
-    yield MyState()
+# class MyState(AppState): ...
 
 
-@rusers.post
-async def create_user(
-    user: User, q: int, r: str
-) -> Resp[Json[User | Order], status.OK]:
-    return User(id=user.id, name=user.name, email=user.email)
+# @asynccontextmanager
+# async def lifespan(app: Lihil[MyState]):
+#     yield MyState()
 
 
-rsubu = rusers.sub("{user_id}")
+# @rusers.post
+# async def create_user(
+#     user: User, q: int, r: str
+# ) -> Resp[Json[User | Order], status.OK]:
+#     return User(id=user.id, name=user.name, email=user.email)
 
 
-@rsubu.get(errors=[UserNotFoundError, UserNotHappyError])
-async def get_user(user_id: str | int) -> Resp[Text, status.OK]:
-    if user_id != "5":
-        raise UserNotFoundError("You can't see me!")
-
-    return "aloha"
+# rsubu = rusers.sub("{user_id}")
 
 
-rprofile = Route("profile/{pid}")
+# @rsubu.get(errors=[UserNotFoundError, UserNotHappyError])
+# async def get_user(user_id: str | int) -> Resp[Text, status.OK]:
+#     if user_id != "5":
+#         raise UserNotFoundError("You can't see me!")
+
+#     return "aloha"
 
 
-class Engine: ...
+# rprofile = Route("profile/{pid}")
 
 
-def get_engine() -> Engine:
-    return Engine()
+# class Engine: ...
 
 
-rprofile.factory(get_engine)
+# def get_engine() -> Engine:
+#     return Engine()
 
 
-@rprofile.post
-async def profile(pid: str, q: int, user: User, engine: Engine) -> User:
-    return User(id=user.id, name=user.name, email=user.email)
+# rprofile.factory(get_engine)
 
 
-rstream = Route("stream")
+# @rprofile.post
+# async def profile(pid: str, q: int, user: User, engine: Engine) -> User:
+#     return User(id=user.id, name=user.name, email=user.email)
 
 
-@rstream.get
-async def stream() -> Stream[str]:
-    const = ["hello", "world"]
-    for c in const:
-        yield c
+# rstream = Route("stream")
 
 
-lhl = Lihil(lifespan=lifespan)
-lhl.include_routes(rusers, rprofile, rstream)
+# @rstream.get
+# async def stream() -> Stream[str]:
+#     const = ["hello", "world"]
+#     for c in const:
+#         yield c
 
 
-lhl.static("/ping", "pong")
+# lhl = Lihil(lifespan=lifespan)
+# lhl.include_routes(rusers, rprofile, rstream)
 
 
-@lhl.post(errors=VioletsAreBlue)
-async def roses_are_red():
-    raise VioletsAreBlue("I am a pythonista")
+# lhl.static("/ping", "pong")
 
 
-# if __name__ == "__main__":
-#     run(lhl)
+# @lhl.post(errors=VioletsAreBlue)
+# async def roses_are_red():
+#     raise VioletsAreBlue("I am a pythonista")
+
+
+# # if __name__ == "__main__":
+# #     run(lhl)
+
+
+from lihil import Lihil
+
+lhl = Lihil()
+
+
+@lhl.get
+async def hello():
+    return "world"