+++ CHANGELOG.md
@@ -27,3 +27,14 @@ This is the very first version of lihil, but we already have a working version t
 
 - fix a bug where `problem.__name__` is used for search param instead of `problem.__problem_type__`
 - no longer import our experimental server by default
+
+## version 0.1.3
+
+### Fix
+
+- fix a bug where if lifespan is not provided, callstack won't be built
+- remove `loguru.logger` as dependency.
+
+### Improvements
+
+- `static` now works with `uvicorn`